postgresql/queries/requirements.txt
argparse-logging==2020.11.26
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
cycler==0.10.0
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
packaging==20.4
pandas==1.1.4
Pillow==8.0.1
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.9.0
pyarrow==2.0.0
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
six==1.15.0
toml==0.10.2
urllib3==1.26.2
zipp==3.4.0
